+++ lib/containers/Overview.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { Header } from '@bpanel/bpanel-ui';
 import { WalletsList, PendingMultiparty, RecentTxList } from '../components';
 
-import { overview } from '../maps';
+import { overview } from '../mappings';
 
 class Overview extends PureComponent {
   constructor(props) {
@@ -24,12 +24,12 @@ class Overview extends PureComponent {
   }
 
   render() {
-    const { walletsList } = this.props;
+    const { walletsList, pendingMultiparty } = this.props;
     return (
       <div>
         <Header type="h2">Overview</Header>
         <WalletsList walletsList={walletsList} />
-        <PendingMultiparty />
+        <PendingMultiparty pendingMultiparty={pendingMultiparty} />
         <RecentTxList />
       </div>
     );
