+++ lib/mappings/README.md
@@ -1,0 +1,1 @@
+Modularizing `mapStateToProps` and `mapDispatchToProps` in this directory.
